Crate popsicle

popsicle: A rust library for private set intersection

popsicle is a rust library that provides implementations for private set intersection (PSI) protocols.

popsicle currently implements the following PSI protocols:

The Pinkas-Schneider-Zohner semi-honest two-party PSI protocol based on the oblivious PRF of Kolesnikov-Kumaresan-Rosulek-Trieu.

The Pinkas-Schnedier-Tkachenko-Yanai semi-honest two-party PSI protocol based on the oblivious programmable PRF of Kolesnikov-Matania-Pinkas-Rosulek-Trieu.

The Kolesnikov-Matania-Pinkas-Rosulek-Trieu semi-honest multi-party PSI protocol.

popsicle should be considered unstable and under active development until version 1.0 is released

Documentation

The documentation can be found here: https://galoisinc.github.io/swanky/popsicle/

Building

Use cargo build to build, cargo test to run the test suite, and cargo bench to benchmark the various protocols.

popsicle also supports the following features:

nightly: Use nightly features from rust and the underlying libraries.

unstable: Enable unstable components of popsicle.

License

MIT License

Authors

Alex J. Malozemoff amaloz@galois.com

Brent Carmer bcarmer@galois.com

Acknowledgments

This material is based upon work supported by the ARO and DARPA under Contract No. W911NF-15-C-0227 and by DARPA and SSC Pacific under Contract No. N66001-15-C-4070.

Any opinions, findings and conclusions or recommendations expressed in this material are those of the author(s) and do not necessarily reflect the views of the ARO, SSC Pacific, and DARPA.

Copyright © 2019 Galois, Inc.

Modules

kmprt
Implementation of the "Kolesnikov-Matania-Pinkas-Rosulek-Trieu" multi-party private set intersection protocol (cf. https://eprint.iacr.org/2017/799.pdf).

psty
Implementation of the Pinkas-Schneider-Tkachenko-Yanai "extended" private set intersection protocol (cf. https://eprint.iacr.org/2019/241).

psz
Implementation of the Pinkas-Schneider-Zohner private set intersection protocol (cf. https://eprint.iacr.org/2014/447) as specified by Kolesnikov-Kumaresan-Rosulek-Trieu (cf. https://eprint.iacr.org/2016/799).

Enums

Error
Errors produced by the private set intersection protocols.

Type Definitions

ExtendedReceiver
Extended private set intersection receiver.

ExtendedSender
Extended private set intersection sender.

MultiPartyReceiver
Multi-party private set intersection receiver.

MultiPartySender
Multi-party private set intersection sender.

Receiver
Private set intersection receiver.

Sender
Private set intersection sender.